核心模块采用内存池模式包括消息传递、套接字收发缓存,编译通过

+++ include/types.h
@@ -10,7 +10,7 @@ typedef unsigned int UID;
 
 typedef unsigned short TIMEERID;
 typedef unsigned short SERVICEINDEX;
-typedef unsigned short DATASIZE;
+typedef unsigned DATASIZE;
 #define MAX_SERVICE_NUM 0xFFFF
 #define INVALID_SERIVCE_INDEX 0
